GeekBrains: add API versioning, redoc/swagger docs
