+++ requirements.txt
@@ -5,27 +5,44 @@ certifi==2019.11.28
 chardet==3.0.4
 colorama==0.4.3
 contextlib2==0.6.0
+coreapi==2.3.3
+coreschema==0.0.4
 distlib==0.3.0
 distro==1.4.0
 Django==3.2.2
 django-cors-headers==3.7.0
 django-filter==2.4.0
 djangorestframework==3.12.4
+djangorestframework-camel-case==1.2.0
+djangorestframework-simplejwt==4.7.1
+drf-yasg==1.20.0
+Faker==8.6.0
 html5lib==1.0.1
 idna==2.8
+inflection==0.5.1
 ipaddr==2.2.0
+itypes==1.2.0
+Jinja2==3.0.1
 lockfile==0.12.2
 Markdown==3.3.4
+MarkupSafe==2.0.1
+mixer==7.1.2
 msgpack==0.6.2
 packaging==20.3
 pep517==0.8.2
 progress==1.5
+PyJWT==2.1.0
 pyparsing==2.4.6
+python-dateutil==2.8.1
 pytoml==0.1.21
 pytz==2021.1
 requests==2.22.0
 retrying==1.3.3
+ruamel.yaml==0.17.9
+ruamel.yaml.clib==0.2.2
 six==1.14.0
 sqlparse==0.4.1
+text-unidecode==1.3
+uritemplate==3.0.1
 urllib3==1.25.8
 webencodings==0.5.1
